+++ src/ui/misc.hpp
@@ -1,13 +1,14 @@
 #ifndef MISC_HEADER
 #define MISC_HEADER
 
+#include "./colors.hpp"
 #include <cctype>
 #include <ftxui/component/captured_mouse.hpp>
 #include <ftxui/component/component.hpp>
 #include <ftxui/component/component_base.hpp>
+#include <ftxui/component/event.hpp>
 #include <ftxui/component/screen_interactive.hpp>
 #include <ftxui/dom/elements.hpp>
-#include <ftxui/screen/color.hpp>
 #include <string>
 #include <vector>
 
@@ -61,7 +62,22 @@ std::vector<std::string> formatLyrics(const std::string& lyrics)
   return lines;
 }
 
-std::string charToStr(char ch) { return std::string(1, static_cast<char>(ch)); }
+std::string charToStr(char ch)
+{
+  switch (ch)
+  {
+    case '\t':
+      return "Tab";
+    case ' ':
+      return "Space";
+    case '\n':
+      return "Enter";
+    case 27:
+      return "Escape"; // ASCII value for the Escape key
+    default:
+      return std::string(1, static_cast<char>(ch));
+  }
+}
 
 std::string FormatTime(int seconds)
 {
@@ -80,6 +96,16 @@ std::string format_song_info(const auto& disc_number, const auto& track_number,
 {
   return std::to_string(disc_number) + "-" + std::to_string(track_number) + ": " + song_name +
          " ( " + FormatTime(duration) + ")";
+}
+
+ftxui::Decorator getTrueColor(TrueColors::Color color)
+{
+  return ftxui::color(TrueColors::GetColor(color));
+}
+
+ftxui::Decorator getTrueBGColor(TrueColors::Color color)
+{
+  return ftxui::bgcolor(TrueColors::GetColor(color));
 }
 
 #endif